add bugfix, change patch location
add bugfix
move ui patch function to client folder
