+++ Contents/mods/!pz_debugTools/media/lua/shared/debugShared/Vehicles.lua
@@ -34,7 +34,7 @@ function Vehicles.getLast(scriptName)
     local test
     for i = allVehicles:size() - 1, 0, -1 do
         test = allVehicles:get(i)
-        if scriptName == test:getScriptName() then
+        if not scriptName or scriptName == test:getScriptName() then
             return test
         end
     end
@@ -46,7 +46,7 @@ function Vehicles.reload(scriptName)
     reloadVehicles()
     reloadVehicleTextures(scriptName)
 
-    if vehicle then vehicle:scriptReloaded() end
+    if vehicle ~= nil then vehicle:scriptReloaded() end
     --if vehicle then Vehicles.remove(vehicle) end
     --Vehicles.add(scriptName)
 end
@@ -95,43 +95,11 @@ function Vehicles.testAllParts(v,test)
         end
     end
 
-    local printTable = {[1]=string.format("\n/*** %s Parts ***/",vehicle:getScriptName() or "")}
+    local printTable = {[1]=string.format("\n/*** %s Parts ***/",vehicle:getScriptName() or "unscripted")}
     for i = 0, vehicle:getPartCount() - 1 do
         printTable[i+2] = test(i,vehicle:getPartByIndex(i))
     end
     print(table.concat(printTable))
-end
-
-function Vehicles.patchISSpawnVehicleUI(ISSpawnVehicleUI)
-    local function removeAll() Vehicles.removeAll() end
-    local function reload(ui,button) Vehicles.reload(ui:getVehicle()) end
-    local function addButton(ui,x,y,w,h,text,target,fn)
-        local button = ISButton:new(x,y,w,h,text,target,fn)
-        button.anchorTop = false
-        button.anchorBottom = true
-        button.borderColor = {r=1, g=1, b=1, a=0.1}
-        button:initialise()
-        ui:addChild(button)
-        return button
-    end
-
-    local initialise = ISSpawnVehicleUI.initialise
-    ISSpawnVehicleUI.initialise = function(self)
-        initialise(self)
-
-        local height = self:getHeight()
-
-        self:setHeight(height+35)
-        for i,v in ipairs({ self.getKey,self.repair }) do --self.spawn,self.close
-            v:setY(v:getY()-35)
-        end
-
-        local x,y,w,h = 10,height - 65,80,25
-        local x1 = 110
-        self.zxReload = addButton(self, x, y, w, h, getText("Reload"), self, reload)
-        self.zxRemoveAll = addButton(self,x1, y, w, h, getText("Remove All"), self, removeAll)
-
-    end
 end
 
 debug.Vehicles = Vehicles
